server/requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
flask>=3.0.0
flask-cors>=4.0.0
python-dateutil>=2.8.2
